# src/backend/doc/swagger.yaml
swagger: "2.0"
info:
  description: "TestBuDDy is a test management, requirement management, automation friendly tool."
  version: "1.0.0"
  title: "TestBuDDy"
  contact:
    email: "barbora.blozonova@gmail.com"
  license:
    name: "MIT licence"
host: "127.0.0.1:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /users/{user_id}:
    get:
      tags:
      - "User and role management"
      summary: "Get user (by set path ID)"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      responses:
        404:
          description: "User not found."
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/User"
          description: "Operation successful. Returns user in JSON response."
    put:
      tags:
      - "User and role management"
      summary: "Update user info"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "user info"
        in: "body"
        description: "User's information"
        required: true
        schema:
          type: "object"
          properties:
            username:
              type: "string"
            password:
              type: "string"
            name:
              type: "string"
            surname:
              type: "string"
      responses:
        200:
          description: "Operation successful. Returns updated user in JSON response."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/User"
        400:
          description: "Invalid body."
        404:
          description: "User not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "User and role management"
      summary: "Delete user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "string"
        required: true
      responses:
        200:
          description: "Operation successful."
        404:
          description: "User not found."
        500:
          description: "Operation failed."
  /users/{user_id}/assign-project:
    post:
      tags:
      - "User and role management"
      summary: "Assign project to user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "project info"
        in: "body"
        description: "Project ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            project_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or project not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "User and role management"
      summary: "Remove assigned project from user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "project info"
        in: "body"
        description: "Project ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            project_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or project not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
  /users/{user_id}/assign-role:
    post:
      tags:
      - "User and role management"
      summary: "Assign role to user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "role info"
        in: "body"
        description: "Role ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            role_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or role not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "User and role management"
      summary: "Remove assigned role from user."
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "role info"
        in: "body"
        description: "Role ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            role_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or role not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
  /users/{user_id}/assign-testcase:
    post:
      tags:
      - "User and role management"
      summary: "Assign test case to user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "test case info"
        in: "body"
        description: "Test case ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            testcase_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or test case not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "User and role management"
      summary: "Remove assigned role from user"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "user_id"
        in: "path"
        type: "integer"
        required: true
      - name: "test case info"
        in: "body"
        description: "Test case ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            testcase_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful."
        400:
          description: "Invalid body."
        404:
          description: "User or test case not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
...
  /init-users:
    post:
      tags:
      - "Demo Initialization"
      summary: "Initialize users and roles."
      description: "(tables: user, role etc.)"
      produces:
      - "text/html"
      responses:
        200:
          description: "Operation successful. Test users initialized."
        500:
          description: "Operation failed."
...
definitions:
  Role:
    type: "object"
    required:
    - id
    - name
    properties:
      id:
        type: "integer"
      name:
        type: "string"
  User:
    type: "object"
    required:
    - id
    - name
    - roles
    properties:
      id:
        type: "integer"
      username:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      roles:
        type: "array"
        items:
          $ref: "#/definitions/Role"
      assigned_projects:
        type: "array"
        items:
          type: "integer"
...